refactor(app): extract prompt-input commands+mode into commands-mode.ts
Slice 2 of the prompt-input.tsx slimming line. Pure extraction, no
behavior/DOM/aria/copy/storage-key change.

Moves the file-pick action, mode switch, and the command.register(
"prompt-input", ...) registration into createPromptCommandsAndMode(deps),
returning only { pick } (setMode and the mode keybinds have no external
consumers). The factory is invoked synchronously at the cluster's original
position so command.register's onCleanup stays bound to the component owner
and registration order is unchanged.

addPickedPaths is produced by createPromptAttachments later in the file, so
it is injected as the accessor thunk () => addPickedPaths and resolved as
addPickedPaths() inside pick — preserving the pre-existing forward-reference
closure timing. editorRef and the file-input click are injected as getters/
callbacks. The only textual changes are those three injection seams.

prompt-input.tsx 637 -> 601.
